+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-# Version 1.9.0
+# Version 1.10.0
 
-* Added support for expandable sections.
+* New `[SLEntityViewControllerSection sectionWithDataSource:delegate:]` allows custom UITableViewDataSources and UITableViewDelegates.
 
 # Version 1.9.1
 
 * Removing UIViewControllerRestoration, UIDataSourceModelAssociation in SLEntityViewController.
+
+# Version 1.9.0
+
+* Added support for expandable sections.